Use doubles, this allows for partials, eg, the nether.

+++ src/main/java/com/me4502/Passthrough/config/PassthroughConfiguration.java
@@ -73,7 +73,7 @@ public class PassthroughConfiguration {
 
                 WorldProperties entry = new WorldProperties(key);
 
-                entry.setBlockModifier(worldProperty.getInt("block-modifier", 1));
+                entry.setBlockModifier(worldProperty.getDouble("block-modifier", 1));
 
                 worldProperties.add(entry);
             }
